allow timestamp on template entities
the items types are missung ui but they'll get it after nico-verticality
is merged
now that templates have comments we can move the sign/bloxberg/timestamp
methods up to homogeneize it all

Note: the sql is piggybacking the schema 179 thas was freshly merged
