+++ src/classes/ItemsTypesSqlBuilder.php
@@ -43,7 +43,10 @@ final class ItemsTypesSqlBuilder extends EntitySqlBuilder
             entity.lockedby,
             entity.locked_at,
             entity.metadata,
-            entity.state';
+            entity.state,
+            entity.timestamped,
+            entity.timestamped_at,
+            entity.timestampedby';
     }
 
     #[Override]
