+++ tests/unit/models/TemplatesTest.php
@@ -52,9 +52,4 @@ class TemplatesTest extends \PHPUnit\Framework\TestCase
         $this->Templates->setId(1);
         $this->assertTrue($this->Templates->destroy());
     }
-
-    public function testGetTimestamperFullname(): void
-    {
-        $this->assertEquals('', $this->Templates->getTimestamperFullname());
-    }
 }
